RAG-lambda/requirements.txt
requests
beautifulsoup4
pymupdf
python-dotenv
psycopg2-binary
urllib3
openai